% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_Konto.R
\name{adjust_Konto}
\alias{adjust_Konto}
\title{helps to generate correctly sized length of Sachkonten an 
  Personenkonten by adjusting string size for Mapping-Process}
\usage{
adjust_Konto(x, length, pad)
}
\arguments{
\item{x}{numer to adjust}

\item{length}{string length}

\item{pad}{fill numeric on right side with ...}
}
\value{
numeric
}
\description{
adjust_Konto
}
