% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASSISTant.R, R/utilities.R
\name{scalarInRange}
\alias{scalarInRange}
\alias{numberInRange}
\alias{scalarIntegerInRange}
\alias{integerInRange}
\alias{den.vs}
\title{Is a scalar quantity is a specified range?}
\usage{
scalarInRange(x, low = -Inf, high = Inf)

numberInRange(x, low = -Inf, high = Inf)

scalarIntegerInRange(x, low = -Inf, high = Inf)

integerInRange(x, low = -Inf, high = Inf)

den.vs(v, i, mu.prime, Sigma.prime, fut)
}
\arguments{
\item{low}{the lower bound, default \code{-Inf}}

\item{high}{the upper bound, default \code{Inf}}

\item{v}{the value of the statistic}

\item{i}{the subgroup}

\item{mu.prime}{the conditional mean vector of the distribution of
length \eqn{J - 1}; needs to be multipled by the conditional
value, the parameter \code{v}.}

\item{Sigma.prime}{the conditional covariance matrix of dimension
\eqn{J-1} by \eqn{J-1}}

\item{fut}{the futility boundary, which is \eqn{\tilde{b}}{btilde}
for stages 1 and 2, but \eqn{c} for stage 3}
}
\value{
\code{TRUE} or \code{FALSE}

\code{TRUE} or \code{FALSE}

\code{TRUE} or \code{FALSE}

\code{TRUE} or \code{FALSE}

the conditional probability
}
\description{
Check if the argument is a scalar within specified range

Check if the argument is within specified range

Check if the argument is a scalar integer within specified range

Check if the argument is an integer vector within specified range

The computation involves a \eqn{J-1} multivariate normal integral
of the conditional density of the \eqn{i}-th subgroup statistic
given that it was maximal among all subgroups:
}
\details{
\deqn{
\phi_i(v)(\int_0^v\int_0^v\ldots
\int_0^{\tilde{b}} \phi_v(z_{-i})dz_{-i})
}

where \eqn{z_{-i}} denotes all subgroups other than \eqn{i}.
}
\keyword{internal}
