% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfgo_badge.R
\name{cfgo_badge}
\alias{cfgo_badge}
\title{Use lifecycle badges in your documentation}
\usage{
cfgo_badge(stage)
}
\arguments{
\item{stage}{choose one of "experimental", "questioning ", "stable" and 
"deprecated".}
}
\description{
Use badges in your function documentation to explain the current status
of the function.You can choose between four different badges: "experimental", 
"questioning ", "stable" and "deprecated". When to use which one while 
following the cfgo style is here explained \url{https://style.cfgo.de/lifecycle}.
}
\details{
Before you can use `cfgo_badge` in your source package you have to run 
`usethis::use_lifecycle()` in the console. This will create the directory
`figures/` in  `man/`. `figures` containes badge images in form of `.svg` 
files. Then you can create badges in your functions. If you want to add
for example the badge "questioning" in your documentation, just copy and paste
\cr\cr
`\\Sexpr[results=rd, stage=render]{assistant::cfgo_badge("questioning")}`
\cr\cr
into your documentation. To check if it works, you have to rebuild your 
source package (Ctrl+Shit+B). For the other badge options just replace 
`questioning` in the code.
}
\section{How to use in function documentation}{

 1. run `usethis::use_lifecycle()`\cr
 2. Use the following script for the badges:
  
 #' Title\cr
 #' \cr
 #' Here ist the discription\cr
 #' \\Sexpr[results=rd, stage=render]assistant::cfgo_badge("questioning")\cr
 #'\cr
 #' @param variables\cr
 #'\cr
 #' @return\cr
 #' @export\cr
 #'\cr
 #' @section Questioning:\cr
 #' Please describe briefly the questioned part. For more detailed
 information refer to the github issue.\cr
 #' \cr
 my_function <- function(variables) \{
 
 \}
}

\examples{
assistant::cfgo_badge("questioning")
 
}
