% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_document_reload.R
\name{clear_document_reload}
\alias{clear_document_reload}
\title{clear, document and reload package}
\usage{
clear_document_reload()
}
\description{
clear_document_reload
}
\details{
clear namespace by detaching all loaded packages document package 
  and reload development version.
}
\examples{
  \dontrun{
    document_and_reload()
  }
}
\seealso{
\link[golem]{document_and_reload}
}
