% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{colNamesForStage}
\alias{colNamesForStage}
\title{Return a vector of column names for statistics for a given stage}
\usage{
colNamesForStage(stage, J)
}
\arguments{
\item{stage}{the trial stage (1 to 3 inclusive).}

\item{J}{the number of subgroups}
}
\value{
a character vector of the column names
}
\description{
Return a vector of column names for statistics for a given stage
}
