% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{computeMeanAndSD}
\alias{computeMeanAndSD}
\title{Compute the mean and sd of a discrete Rankin distribution}
\usage{
computeMeanAndSD(probVec = rep(1, 7))
}
\arguments{
\item{probVec}{a probability vector of length 7 (implicit support on \code{0:6})}
}
\value{
a named vector of \code{mean} and \code{sd}
}
\description{
Compute the mean and sd of a discrete Rankin distribution
}
