% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{conformParameters}
\alias{conformParameters}
\title{Conform designParameters so that weights are turned in to probabilities, the null and control distributions are proper matrices etc.}
\usage{
conformParameters(plist, discreteData = FALSE)
}
\arguments{
\item{plist}{the parameter list}

\item{discreteData}{flag if data is discrete}
}
\value{
the modified parameter list
}
\description{
Conform designParameters so that weights are turned in to probabilities, the null and control distributions are proper matrices etc.
}
