% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generateDiscreteData}
\alias{generateDiscreteData}
\title{A data generation function using a discrete distribution for Rankin
score rather than a normal distribution}
\usage{
generateDiscreteData(prevalence, N, ctlDist, trtDist)
}
\arguments{
\item{prevalence}{a vector of group prevalences (length denoted by J below)}

\item{N}{the sample size to generate}

\item{ctlDist}{a probability vector of length 7 denoting the Rankin Score distribution for control}

\item{trtDist}{a 7 x J probability matrix with each column is the Ranking distribution for the associated group}
}
\value{
a three-column data frame of \code{subGroup}, \code{trt} (0 or 1), and \code{score}
}
\description{
A data generation function using a discrete distribution for Rankin
score rather than a normal distribution
}
\examples{
# Simulate data from a discrete distribution for the Rankin scores,
# which are ordinal integers from 0 to 6 in the following
# simulations. So we define a few scenarios.
library(ASSISTant)
null.uniform <- rep(1, 7L) ## uniform on 7 support points
hourglass <- c(1, 2, 2, 1, 2, 2, 1)
inverted.hourglass <- c(2, 1, 1, 2, 1, 1, 2)
bottom.heavy <- c(2, 2, 2, 1, 1, 1, 1)
bottom.heavier <- c(3, 3, 2, 2, 1, 1, 1)
top.heavy <- c(1, 1, 1, 1, 2, 2, 2)
top.heavier <- c(1, 1, 1, 2, 2, 3, 3)
ctlDist <- null.uniform
trtDist <- cbind(null.uniform, null.uniform, hourglass, hourglass)
generateDiscreteData(prevalence = rep(1, 4), 10, ctlDist, trtDist)
trtDist <- cbind(bottom.heavy, bottom.heavy, top.heavy, top.heavy)
generateDiscreteData(prevalence = rep(1, 4), 10, ctlDist, trtDist)
}
