% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generateNormalData}
\alias{generateNormalData}
\title{A data generation function along the lines of what was used in the Lai, Lavori, Liao paper.
score rather than a normal distribution}
\usage{
generateNormalData(prevalence, N, mean, sd)
}
\arguments{
\item{prevalence}{a vector of group prevalences (length denoted by J below)}

\item{N}{the sample size to generate}

\item{mean}{a 2 x J matrix of means under the null (first row) and alternative for each group}

\item{sd}{a 2 x J matrix of standard deviations under the null (first row) and alternative for each group}
}
\value{
a three-column data frame of \code{subGroup}, \code{trt} (0 or 1), and \code{score}
}
\description{
A data generation function along the lines of what was used in the Lai, Lavori, Liao paper.
score rather than a normal distribution
}
