% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mHP.btilde}
\alias{mHP.btilde}
\title{Compute the futility boundary (modified Haybittle-Peto) for the
first two stages}
\usage{
mHP.btilde(beta, cov.J)
}
\arguments{
\item{beta}{the type II error}

\item{cov.J}{the 3 x 3 covariance matrix}
}
\description{
The futility boundary \eqn{\tilde{b}}{btilde} is computed by
solving (under the alternative)
}
\details{
\deqn{
P(\tilde{Z}_J^1\le\tilde{b} or \tilde{Z}_J^2\le\tilde{b}) = \epsilon\beta }

where the superscripts denote the stage and \eqn{\epsilon} is the
fraction of the type I error (\eqn{\alpha}) spent and \eqn{\beta}
is the type II error. We make use of the joint normal density of
\eqn{Z_{J}} (the overall group) at each of the three stages and the
fact that the \eqn{\tilde{Z_J}} is merely a translation of
\eqn{Z_J}. So here the calculation is based on a mean of zero and
has to be translated during use!
}
\references{
Adaptive Choice of Patient Subgroup for Comparing Two
Treatments by Tze Leung Lai and Philip W. Lavori and Olivia
Yueh-Wen Liao. Contemporary Clinical Trials, Vol. 39, No. 2, pp
191-200
(2014). \url{http://www.sciencedirect.com/science/article/pii/S1551714414001311}
}
