% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASSISTant.R
\name{scalarInRange}
\alias{integerInRange}
\alias{numberInRange}
\alias{scalarInRange}
\alias{scalarIntegerInRange}
\title{Is a scalar quantity is a specified range?}
\usage{
scalarInRange(x, low = -Inf, high = Inf)

numberInRange(x, low = -Inf, high = Inf)

scalarIntegerInRange(x, low = -Inf, high = Inf)

integerInRange(x, low = -Inf, high = Inf)
}
\arguments{
\item{low}{the lower bound, default \code{-Inf}}

\item{high}{the upper bound, default \code{Inf}}

\item{low}{the lower bound, default \code{-Inf}}

\item{high}{the upper bound, default \code{Inf}}

\item{low}{the lower bound, default \code{-Inf}}

\item{high}{the upper bound, default \code{Inf}}

\item{low}{the lower bound, default \code{-Inf}}

\item{high}{the upper bound, default \code{Inf}}
}
\value{
\code{TRUE} or \code{FALSE}

\code{TRUE} or \code{FALSE}

\code{TRUE} or \code{FALSE}

\code{TRUE} or \code{FALSE}
}
\description{
Check if the argument is a scalar within specified range

Check if the argument is within specified range

Check if the argument is a scalar integer within specified range

Check if the argument is an integer vector within specified range
}
\examples{
ASSISTant:::scalarInRange(x = 10, low = 2)  ## TRUE
ASSISTant:::scalarInRange(x = 10, high = 2) ## FALSE

ASSISTant:::numberInRange(x = 2:10, low = 2)  ## TRUE
ASSISTant:::numberInRange(x = 10:15, high = 2) ## FALSE

ASSISTant:::integerInRange(x = 10, low = 2)  ## TRUE
ASSISTant:::integerInRange(x = 10.5) ## FALSE

ASSISTant:::integerInRange(x = 2:10, low = 2)  ## TRUE
ASSISTant:::integerInRange(x = 10:15, high = 2) ## FALSE
ASSISTant:::integerInRange(x = c(0.5, 1.5), high = 2) ## FALSE

}
\keyword{internal}

