% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASSISTDesigns.R
\docType{class}
\name{ASSISTDesignB}
\alias{ASSISTDesignB}
\title{A fixed sample design to compare against the adaptive clinical trial design of Lai, Lavori and Liao.}
\format{An \code{\link{R6Class}} generator object}
\usage{
ASSISTDesignB
}
\description{
\code{ASSISTDesignB} objects are used to design a trial with certain
characteristics provided in the object instantiation method. This design differs from
\code{ASSISTDesign} in only how it computes the critical boundaries, how it performs the interim look,
and what quantities are computed in a trial run.
}
\section{Methods}{


\describe{
  \item{\code{ASSISTDesignB$new(designParameters, trialParameters, generateData)}}{Create a new \code{ASSISTDesign}
        instance object using the parameters specified. }
  \item{\code{getTrueParameters},\code{getTrialParameters},
        \code{getBoundaries}}{Accessor methods for (obvious) object slots}
  \item{\code{print()}}{Print the object in a human readable form}
  \item{\code{computeCriticalValues()}}{Compute the critical boundary value \eqn{c_\alpha}}
  \item{\code{explore(numberOfSimulations = 100, rngSeed = 12345, showProgress = TRUE)}}{Explore the design using
        the specified number of simulations and random number seed. Show progress if so desired.
        Returns a data frame of results}
  \item{\code{analyze(trialHistory = NULL, numberOfSimulations = 25, rngSeed = 12345, showProgress = TRUE)}}{Analyze the design given. Use the \code{trialHistory} if given, else call \code{explore} to
        simulate the design. Return a list of summary quantities}
  \item{\code{summary(analysis)}}{Print the operating characteristics of the design, using the analysis
        result from the \code{analyze} call}
}
}
\references{
Adaptive Choice of Patient Subgroup for Comparing Two
Treatments \url{http://dx.doi.org/10.106/j.cct.2014.09.001}
}
\seealso{
\code{ASSISTDesign} which is a superclass of this object
}
\keyword{datasets}

