% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{computeMeanAndSD}
\alias{computeMeanAndSD}
\title{Compute the mean and sd of a discrete Rankin distribution}
\usage{
computeMeanAndSD(probVec = rep(1, 7L), support = 0L:6L)
}
\arguments{
\item{probVec}{a probability vector of length equal to length of support,
default is uniform}

\item{support}{a vector of support values (default 0:6 for Rankin Scores)}
}
\value{
a named vector of \code{mean} and \code{sd}
}
\description{
Compute the mean and sd of a discrete Rankin distribution
}
