% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{computeMHPBoundaryITT}
\alias{computeMHPBoundaryITT}
\title{Compute the three modified Haybittle-Peto boundaries and effect size}
\usage{
computeMHPBoundaryITT(prevalence, alpha)
}
\arguments{
\item{prevalence}{the vector of prevalences between 0 and 1 summing
to 1. \eqn{J}, the number of groups, is implicitly the length
of this vector and should be at least 2.}

\item{alpha}{the type I error}
}
\value{
a named vector of a single value containing the value for \code{c}
}
\description{
Compute the three modified Haybittle-Peto boundaries and effect size
}
\references{
Adaptive Choice of Patient Subgroup for Comparing Two
Treatments by Tze Leung Lai and Philip W. Lavori and Olivia
Yueh-Wen Liao. Contemporary Clinical Trials, Vol. 39, No. 2, pp
191-200
(2014). \url{http://www.sciencedirect.com/science/article/pii/S1551714414001311}
}
