% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASSISTDesigns.R
\docType{class}
\name{DEFUSE3Design}
\alias{DEFUSE3Design}
\title{The DEFUSE3 design}
\format{An \code{\link{R6Class}} generator object}
\usage{
DEFUSE3Design
}
\description{
\code{DEFUSE3Design} is a slight variant of the the adaptive
clinical trial design of Lai, Lavori and Liao. Simulation is used to compute
the expected maximum sample size and the boundary for early futility is adjusted to
account as well.
}
\section{Methods}{


\describe{
  \item{\code{DEFUSE3Design$new(designParameters, trialParameters, generateData, numberOfSimulations = 5000, rngSeed = 12345, showProgress = TRUE)}}{Create
        a new \code{ASSISTDesign} instance object using the parameters specified. }
  \item{\code{getDesignParameters},\code{getTrialParameters},
        \code{getBoundaries}}{Accessor methods for (obvious) object slots}
  \item{\code{print()}}{Print the object in a human readable form}
  \item{\code{adjustCriticalValues(numberOfSimulations, rngSeed, showProgress)}}{Adjust the critical values
        by performing simulations using the parameters provided}
  \item{\code{computeCriticalValues()}}{Compute the critical boundary value \eqn{c_\alpha}}
  \item{\code{explore(numberOfSimulations = 100, rngSeed = 12345, trueParameters = self$getDesignParameters(). showProgress = TRUE)}}{Explore the design using
        the specified number of simulations and random number seed.  \code{trueParameters} is by default the same
        as \code{designParameters} as would be the case for a Type I error calculation. If changed, would yield power.
        Show progress if so desired. Returns a data frame of results}
  \item{\code{analyze(trialHistory = NULL, numberOfSimulations = 25, rngSeed = 12345, showProgress = TRUE)}}{Analyze
        the design given. Use the \code{trialHistory} if given, else call \code{explore} to
        simulate the design. Return a list of summary quantities}
  \item{\code{summary(analysis)}}{Print the operating characteristics of the design, using the analysis
        result from the \code{analyze} call}
}
}
\references{
Adaptive Choice of Patient Subgroup for Comparing Two
Treatments \url{http://dx.doi.org/10.106/j.cct.2014.09.001}
}
\seealso{
\code{ASSISTDesign} which is a superclass of this object
}
\keyword{datasets}

