% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generateDiscreteData}
\alias{generateDiscreteData}
\title{A data generation function using a discrete distribution for Rankin
score rather than a normal distribution}
\usage{
generateDiscreteData(prevalence, N, ctlDist, trtDist)
}
\arguments{
\item{prevalence}{a vector of group prevalences (length denoted by J below)}

\item{N}{the sample size to generate}

\item{ctlDist}{a probability vector of length 7 denoting the Rankin Score distribution for control}

\item{trtDist}{a 7 x J probability matrix with each column is the Ranking distribution for the associated group}
}
\value{
a three-column data frame of \code{subGroup}, \code{trt} (0 or 1), and \code{score}
}
\description{
A data generation function using a discrete distribution for Rankin
score rather than a normal distribution
}
